+++ config/environment.js
@@ -53,6 +53,11 @@ module.exports = function(environment) {
                         anonymizeIp: true,
                     },
                 },
+                dimensions: {
+                    authenticated: 'dimension1',
+                    resource: 'dimension2',
+                    isPublic: 'dimension3',
+                },
             },
         ],
         PROVIDER_SETTINGS: [{